producer/requirements.txt
kafka-python==2.0.2
yfinance
fastapi 
uvicorn
redis 
